+++ Slim/Routing/RouteCollector.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
 
 namespace Slim\Routing;
 
-use FastRoute\RouteParser;
 use FastRoute\RouteParser\Std as StdParser;
 use InvalidArgumentException;
 use Psr\Container\ContainerInterface;
@@ -22,6 +21,7 @@ use Slim\Interfaces\InvocationStrategyInterface;
 use Slim\Interfaces\RouteCollectorInterface;
 use Slim\Interfaces\RouteGroupInterface;
 use Slim\Interfaces\RouteInterface;
+use Slim\Interfaces\RouteParserInterface;
 
 /**
  * RouteCollector is used to collect routes and route groups
@@ -30,9 +30,7 @@ use Slim\Interfaces\RouteInterface;
 class RouteCollector implements RouteCollectorInterface
 {
     /**
-     * Parser
-     *
-     * @var RouteParser
+     * @var RouteParserInterface
      */
     protected $routeParser;
 
@@ -98,28 +96,28 @@ class RouteCollector implements RouteCollectorInterface
      * @param CallableResolverInterface     $callableResolver
      * @param ContainerInterface|null       $container
      * @param InvocationStrategyInterface   $defaultInvocationStrategy
-     * @param RouteParser                   $parser
+     * @param RouteParserInterface          $routeParser
      */
     public function __construct(
         ResponseFactoryInterface $responseFactory,
         CallableResolverInterface $callableResolver,
         ContainerInterface $container = null,
         InvocationStrategyInterface $defaultInvocationStrategy = null,
-        RouteParser $parser = null
+        RouteParserInterface $routeParser = null
     ) {
         $this->responseFactory = $responseFactory;
         $this->callableResolver = $callableResolver;
         $this->container = $container;
         $this->defaultInvocationStrategy = $defaultInvocationStrategy ?? new RequestResponse();
-        $this->routeParser = $parser ?? new StdParser;
+        $this->routeParser = $routeParser ?? new RouteParser($this);
     }
 
     /**
-     * @return CallableResolverInterface
+     * @return RouteParserInterface
      */
-    public function getCallableResolver(): CallableResolverInterface
+    public function getRouteParser(): RouteParserInterface
     {
-        return $this->callableResolver;
+        return $this->routeParser;
     }
 
     /**
@@ -308,103 +306,5 @@ class RouteCollector implements RouteCollectorInterface
             $this->routeGroups,
             $this->routeCounter
         );
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function relativePathFor(string $name, array $data = [], array $queryParams = []): string
-    {
-        $route = $this->getNamedRoute($name);
-        $pattern = $route->getPattern();
-
-        $segments = [];
-        $segmentName = '';
-
-        /*
-         * $routes is an associative array of expressions representing a route as multiple segments
-         * There is an expression for each optional parameter plus one without the optional parameters
-         * The most specific is last, hence why we reverse the array before iterating over it
-         */
-        $expressions = array_reverse($this->routeParser->parse($pattern));
-        foreach ($expressions as $expression) {
-            foreach ($expression as $segment) {
-                /*
-                 * Each $segment is either a string or an array of strings
-                 * containing optional parameters of an expression
-                 */
-                if (is_string($segment)) {
-                    $segments[] = $segment;
-                    continue;
-                }
-
-                /*
-                 * If we don't have a data element for this segment in the provided $data
-                 * we cancel testing to move onto the next expression with a less specific item
-                 */
-                if (!array_key_exists($segment[0], $data)) {
-                    $segments = [];
-                    $segmentName = $segment[0];
-                    break;
-                }
-
-                $segments[] = $data[$segment[0]];
-            }
-
-            /*
-             * If we get to this logic block we have found all the parameters
-             * for the provided $data which means we don't need to continue testing
-             * less specific expressions
-             */
-            if (!empty($segments)) {
-                break;
-            }
-        }
-
-        if (empty($segments)) {
-            throw new InvalidArgumentException('Missing data for URL segment: ' . $segmentName);
-        }
-
-        $url = implode('', $segments);
-        if ($queryParams) {
-            $url .= '?' . http_build_query($queryParams);
-        }
-
-        return $url;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function pathFor(string $name, array $data = [], array $queryParams = []): string
-    {
-        trigger_error('pathFor() is deprecated. Use urlFor() instead.', E_USER_DEPRECATED);
-        return $this->urlFor($name, $data, $queryParams);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function urlFor(string $name, array $data = [], array $queryParams = []): string
-    {
-        $url = $this->relativePathFor($name, $data, $queryParams);
-
-        if ($this->basePath) {
-            $url = $this->basePath . $url;
-        }
-
-        return $url;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function fullUrlFor(UriInterface $uri, string $routeName, array $data = [], array $queryParams = []): string
-    {
-        $path = $this->urlFor($routeName, $data, $queryParams);
-        $scheme = $uri->getScheme();
-        $authority = $uri->getAuthority();
-        $protocol = ($scheme ? $scheme . ':' : '') . ($authority ? '//' . $authority : '');
-        return $protocol . $path;
     }
 }